parinama/backend/requirements.txt
# ══════════════════════════════════════════════
# PARINAMA — Self-Evolving Prompt Optimization Engine
# Backend Dependencies — 100% FREE, NO PAID APIs
# ══════════════════════════════════════════════

# --- Web Framework ---
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# --- LLM SDKs (ALL FREE) ---
groq==0.4.2
google-generativeai==0.8.0
ollama==0.1.7

# --- Prompt Optimization ---
dspy-ai==2.4.0

# --- Database ---
sqlalchemy==2.0.23
aiosqlite==0.19.0

# --- Data Validation ---
pydantic==2.5.0
pydantic-settings==2.1.0

# --- Environment & Config ---
python-dotenv==1.0.0

# --- Security & Rate Limiting ---
slowapi==0.1.9
bleach==6.1.0

# --- Utilities ---
httpx==0.25.2
tenacity==8.2.3
aiofiles==23.2.1

# --- Async Support ---
anyio==3.7.1
